Refactor and segfault bugfix
I was wanting to use this again on a new Ubuntu system, only to find
that for whatever reason the program segfaulted.

The fix was a simple missing colon in the getopt string, but while I
was looking at that I noticed just how sloppy the code was and cleaned
it up.

For the record, I can't get it working in Ubuntu right now,
as I can get spaceball events fine but I can't make a fake device.
Previously I was using an openSUSE machine, and the old code worked
perfectly fine on there, so I don't know what could be different.
Perhaps I need to install something? IDK.